change the view to the page
